move impact header view to own file
